requirements.txt
faster-whisper>=1.0.0
whisperx>=3.1.1  # Optional: Only needed for speaker diarization (requires HF token)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
pydantic>=2.5.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
click>=8.1.7
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
